+++ CHANGELOG.md
@@ -16,6 +16,11 @@ speaking, considered to be the player's experience.
 - Show the average target resistance of independent towns on the information
   panel when you have workers.
 
+### Fixed
+
+- Fixed target resistance being shown incorrectly when you have workers in the
+  independent town.
+
 ### Changed
 
 - Base the display of the building guide on the Help setting.  Not shown when